requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
gitpython==3.1.44
keyboard==0.13.5
bitsandbytes==0.45.3
llama-cpp-python==0.3.8